Demo Dataset for Dropout in an Online Survey

Simulated demo data set for dropout in a survey.

Format

A data frame with 246 rows and 54 variables (in the order they were presented in the fictional survey).

obs_id

Observation ID

experimental_condition

experimental condition

vi_1

item 1

vi_2

item 2

vi_3

item 3

vi_4

item 4

vi_5

item 5

vi_6

item 6

vi_7

item 7

vi_8

item 8

vi_9

item 9

vi_10

item 10

vi_11

item 11

vi_12

item 12

vi_13

item 13

vi_14

item 14

vi_15

item 15

vi_16

item 16

vi_17

item 17

vi_18

item 18

vi_19

item 19

vi_20

item 20

vi_21

item 21

vi_22

item 22

vi_23

item 23

vi_24

item 24

vi_25

item 25

vi_26

item 26

vi_27

item 27

vi_28

item 28

vi_29

item 29

vi_30

item 30

vi_31

item 31

vi_32

item 32

vi_33

item 33

vi_34

item 34

vi_35

item 35

vi_36

item 36

vi_37

item 37

vi_38

item 38

vi_39

item 39

vi_40

item 40

vi_41

item 41

vi_42

item 42

vi_43

item 43

vi_44

item 44

vi_45

item 45

vi_46

item 46

vi_47

item 47

vi_48

item 48

vi_49

item 49

vi_50

item 50

vi_51

item 51

vi_52

item 52

Source

dropRdemo Demo data for dropout.